feat: #308 Gemini/Custom Search 외부 호출에 서킷브레이커 적용
